Santix

+++ Classes/kernel_attributes.py
@@ -48,15 +48,17 @@ class kernel_attributes:
         except:
             print("Data not found")
         return query,columns
-    def delete_table(self,dat):
+    def delete_table(self,dat,db):
         id=self.selection(dat)
         columns=self.attributes
         try:
             for j in range(len(id)):
-               for i in range(len(columns)):
-                    obj=self.attributesT[columns[i].name]
-                    obj.delete_name(id[j])
-                    obj.delete_uuid(id[j])
+                if self.reference_comprobation(columns,db,id[j],"Delete")==len(columns):
+                    for i in range(len(columns)):
+                        obj=self.attributesT[columns[i].name]
+                        obj.delete_name(id[j])
+                else:
+                    print("Rerror")    
         except:
             print("Data not found")
     def update_table(self,set,dat):
@@ -69,6 +71,11 @@ class kernel_attributes:
                     obj.update_uuid(id[j],set[i+1])
         except:
             print("Data not found")
+    def reference_table(self,attribute,name,atr):
+        try:
+           self.attributesT[attribute].referenced[name]=atr
+        except:
+            print("Attribute error")
 
 #Relational algebra operations
     def selection(self,dat):
@@ -100,6 +107,13 @@ class kernel_attributes:
         else:
             id_f=obj.select_name(dat)
         return id_f
+#Comprobations
+    def reference_comprobation(self,columns,db,dat,type):
+       cont=0
+       for i in range(len(columns)):
+            obj=self.attributesT[columns[i].name]
+            cont=obj.reference_comprobation(dat,db,cont,type)
+       return cont
